chore(quickstart): split storybook stories into per-variant folders
Scenes-App/Quickstart now has three folders: 'Full page (test)',
'Simplified (test2)', and the existing 'Product card'. Shared scenario
mocks and installation-state decorators move to quickstartStoryScenarios.
The test2 folder drops test-skip: without the publications rail there is
no external artwork, so its five states get visual regression coverage.
